delete mocks, refactor Logo, addSpinner to favorites
