+++ src/components/logo/logo.tsx
@@ -1,26 +1,26 @@
-import { Link } from 'react-router-dom';
+import cn from 'classnames';
+import { Link, useLocation } from 'react-router-dom';
 import { APP_ROUTE } from '../../const';
 
 type LogoProps = {
-  linkClassName: string;
   isFooter?: boolean;
 };
 
-export default function Logo({ linkClassName, isFooter }: LogoProps) {
+export default function Logo({ isFooter }: LogoProps) {
+  //TODO: убрать в константы
   const imgWidth = isFooter ? 64 : 81;
   const imgHeight = isFooter ? 33 : 41;
-  return linkClassName ? (
-    <div>
-      <img
-        className="header__logo"
-        src="img/logo.svg"
-        alt="6 cities logo"
-        width={imgWidth}
-        height={imgHeight}
-      />
-    </div>
-  ) : (
-    <Link to={APP_ROUTE.ROOT} className={`header__logo-link${linkClassName}`}>
+
+  const { pathname } = useLocation();
+  const isMainPage = pathname === APP_ROUTE.ROOT;
+  return (
+    <Link
+      to={APP_ROUTE.ROOT}
+      className={cn(
+        'header__logo-link',
+        isMainPage && 'header__logo-link--active'
+      )}
+    >
       <img
         className="header__logo"
         src="img/logo.svg"
