+++ src/components/offer-container/offer-container.tsx
@@ -3,6 +3,7 @@ import {
   getBadroomsString,
   getAdultsString,
 } from '../../helpers/getPluralString';
+import { useAuthCheck } from '../../hooks/use-auth-check';
 import { OfferDetail, Review } from '../../types';
 import Avatar from '../avatar/avatar';
 import FavoriteButton from '../favorite-button/favorite-button';
@@ -16,15 +17,12 @@ type OfferContainerProps = {
   currentOffer: OfferDetail;
 };
 
-//TODO: replace with actual logic
-const isAuth = true;
-
 export default function OfferContainer({
   reviews,
   currentOffer,
 }: OfferContainerProps) {
   const {
-    // id,
+    id,
     isFavorite,
     isPremium,
     title,
@@ -37,13 +35,14 @@ export default function OfferContainer({
     host,
     description,
   } = currentOffer;
+  const isAuth = useAuthCheck();
   return (
     <div className="offer__container container">
       <div className="offer__wrapper">
         {isPremium && <PremiumBadge isOfferDetail />}
         <div className="offer__name-wrapper">
           <h1 className="offer__name">{title}</h1>
-          <FavoriteButton className="offer" isFavorite={isFavorite} />
+          <FavoriteButton offerId={id} classNamePrefix="offer" isFavorite={isFavorite} />
         </div>
         <OfferRating rating={rating} isOfferDetail />
         <ul className="offer__features">
