--- src/mocks/comments.ts
@@ -1,50 +1,0 @@
-export const COMMENTS = [
-  {
-    id: '6f6becc6-b05b-4954-a165-6383747770b7',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Oliver Conner',
-      avatarUrl: 'img/avatar-max.jpg',
-      isPro: false,
-    },
-    comment:
-      'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    rating: 4,
-  },
-  {
-    id: 'test1',
-    date: '2018-05-08T14:13:56.569Z',
-    user: {
-      name: 'Oliver Conner',
-      avatarUrl: 'img/avatar-max.jpg',
-      isPro: false,
-    },
-    comment:
-      'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    rating: 4,
-  },
-  {
-    id: 'test2',
-    date: '2017-05-08T14:13:56.569Z',
-    user: {
-      name: 'Oliver Conner',
-      avatarUrl: 'img/avatar-max.jpg',
-      isPro: false,
-    },
-    comment:
-      'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    rating: 4,
-  },
-  {
-    id: '6f6becc6-b05b-4954-a165-6383747770b8',
-    date: '2016-05-08T14:13:56.569Z',
-    user: {
-      name: 'Oliver Conner',
-      avatarUrl: 'img/avatar-max.jpg',
-      isPro: true,
-    },
-    comment:
-      'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    rating: 4,
-  },
-];
