--- src/mocks/favorites.ts
@@ -1,117 +1,0 @@
-export const FAV_OFFERS = [
-  {
-    id: 'd2afa623-51ca-443c-9ac7-4942725c3970',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'hotel',
-    price: 486,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/18.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.868610000000004,
-      longitude: 2.342499,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: false,
-    rating: 3.6,
-  },
-  {
-    id: '36017de2-a30d-424e-bf26-30ff41d58765',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'house',
-    price: 425,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.214402,
-      longitude: 6.764314000000001,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: true,
-    rating: 1.6,
-  },
-  {
-    id: '9083425e-e7f5-41f8-9fa4-4ba96f85925d',
-    title: 'Loft Studio in the Central Area',
-    type: 'hotel',
-    price: 355,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/11.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.225402,
-      longitude: 6.784314,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: false,
-    rating: 4.4,
-  },
-  {
-    id: '286c107f-5ed9-4a12-ba9e-e16b5cc4836b',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'apartment',
-    price: 263,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36554,
-      longitude: 4.911976,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: true,
-    rating: 1.7,
-  },
-  {
-    id: '00cd8b57-6c63-4b22-8a81-27bdafe72930',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'house',
-    price: 377,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/4.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.385540000000006,
-      longitude: 4.902976,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: false,
-    rating: 4.9,
-  },
-];
