--- src/mocks/offers.ts
@@ -1,2818 +1,0 @@
-export const OFFERS = [
-  {
-    id: 'd2afa623-51ca-443c-9ac7-4942725c3970',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'hotel',
-    price: 486,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/18.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.868610000000004,
-      longitude: 2.342499,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: false,
-    rating: 3.6,
-  },
-  {
-    id: '43378ad2-368f-471a-bdd6-800b10096182',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'apartment',
-    price: 300,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.858610000000006,
-      longitude: 2.330499,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: false,
-    rating: 1.8,
-  },
-  {
-    id: '0aa04fb2-db56-4efb-bd3a-d585acb0f3f8',
-    title: 'Perfectly located Castro',
-    type: 'house',
-    price: 328,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.834610000000005,
-      longitude: 2.335499,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: false,
-    rating: 1,
-  },
-  {
-    id: 'd371db05-3a9a-49c6-a5d3-df424d91af14',
-    title: 'House in countryside',
-    type: 'room',
-    price: 171,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/20.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.85761,
-      longitude: 2.358499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2,
-  },
-  {
-    id: '3861ed93-f196-4862-8e0c-4906eb2f7130',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'room',
-    price: 220,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.87561,
-      longitude: 2.375499,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: true,
-    rating: 3.5,
-  },
-  {
-    id: '30923e2b-191b-4ce0-b884-17fb84da6d71',
-    title: 'Perfectly located Castro',
-    type: 'hotel',
-    price: 107,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/18.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.87961000000001,
-      longitude: 2.353499,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: false,
-    rating: 1.3,
-  },
-  {
-    id: '54da406b-10ba-4a18-9888-6b636f46926f',
-    title: 'Wood and stone place',
-    type: 'apartment',
-    price: 304,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.834610000000005,
-      longitude: 2.364499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.6,
-  },
-  {
-    id: 'fe691f66-5e69-485f-bf1a-17f685491584',
-    title: 'Waterfront with extraordinary view',
-    type: 'apartment',
-    price: 281,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.837610000000005,
-      longitude: 2.3454990000000002,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.5,
-  },
-  {
-    id: '00f92ede-9e94-477d-8449-ed098b431411',
-    title: 'Waterfront with extraordinary view',
-    type: 'apartment',
-    price: 174,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.84761,
-      longitude: 2.356499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2,
-  },
-  {
-    id: '8a17e693-b190-477f-bec3-6ae020599516',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'room',
-    price: 265,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.862610000000004,
-      longitude: 2.369499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.5,
-  },
-  {
-    id: 'de89cdb4-c3bc-480d-9b3b-905a430eda29',
-    title: 'Loft Studio in the Central Area',
-    type: 'house',
-    price: 144,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.83861,
-      longitude: 2.350499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2,
-  },
-  {
-    id: 'a5bcf6eb-2055-45cc-b344-b87b50dcb558',
-    title: 'The house among olive ',
-    type: 'house',
-    price: 235,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/7.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.861610000000006,
-      longitude: 2.340499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.7,
-  },
-  {
-    id: '0d753035-6889-4933-ae8c-e43e3d9c0f2d',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'room',
-    price: 294,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.87861,
-      longitude: 2.357499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 5,
-  },
-  {
-    id: 'a77440bb-95fa-40b1-ae0d-87a2b09ea2e1',
-    title: 'Tile House',
-    type: 'room',
-    price: 111,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.877610000000004,
-      longitude: 2.333499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.2,
-  },
-  {
-    id: '4fdee574-25b5-4166-a4e8-ca1d5b9df6f8',
-    title: 'House in countryside',
-    type: 'house',
-    price: 663,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/6.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.83961,
-      longitude: 2.342499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.8,
-  },
-  {
-    id: 'ef5b5b71-f057-4b89-beeb-6870e4ebdc82',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'apartment',
-    price: 333,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/12.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.865610000000004,
-      longitude: 2.350499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.8,
-  },
-  {
-    id: 'e7ca6443-3c00-4dbd-91aa-c23fc291bc11',
-    title: 'Penthouse, 4-5 rooms + 5 balconies',
-    type: 'room',
-    price: 128,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/2.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.846610000000005,
-      longitude: 2.374499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.2,
-  },
-  {
-    id: 'bd80f02c-ee35-466f-9bb5-fce956d8b694',
-    title: 'Waterfront with extraordinary view',
-    type: 'house',
-    price: 113,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/5.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.837610000000005,
-      longitude: 2.364499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.7,
-  },
-  {
-    id: '45fb6485-348f-4d38-8593-3c0ac05827cc',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'apartment',
-    price: 126,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/9.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.843610000000005,
-      longitude: 2.338499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.8,
-  },
-  {
-    id: 'cd9d8137-551f-442c-a312-e7d661ff91f4',
-    title: 'Loft Studio in the Central Area',
-    type: 'house',
-    price: 144,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.84461,
-      longitude: 2.374499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.6,
-  },
-  {
-    id: 'f8f6e1d2-c99d-4141-9721-b48cc46b8827',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'apartment',
-    price: 243,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/11.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.950361,
-      longitude: 6.961974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.4,
-  },
-  {
-    id: 'f970240c-47f4-49ff-9dc2-e1d52b4bb2ce',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'apartment',
-    price: 291,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/18.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.932361,
-      longitude: 6.937974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.4,
-  },
-  {
-    id: '0b6758f3-871f-410b-a7c4-ea858fcd84f6',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'house',
-    price: 171,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.934361,
-      longitude: 6.943974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2,
-  },
-  {
-    id: '8c612a18-38ac-406a-a303-e8347da2cfd4',
-    title: 'The house among olive ',
-    type: 'apartment',
-    price: 286,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/11.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.947361,
-      longitude: 6.9799739999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.1,
-  },
-  {
-    id: '00ddee2c-e8e2-4904-a248-8983d0170464',
-    title: 'Canal View Prinsengracht',
-    type: 'hotel',
-    price: 449,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.960361,
-      longitude: 6.967974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.7,
-  },
-  {
-    id: 'a0484781-6af2-493c-904a-a449a2364a72',
-    title: 'Loft Studio in the Central Area',
-    type: 'apartment',
-    price: 489,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/7.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.941361,
-      longitude: 6.956974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.3,
-  },
-  {
-    id: 'ecef6c17-af14-4eb0-85a1-22b49221a093',
-    title: 'Penthouse, 4-5 rooms + 5 balconies',
-    type: 'room',
-    price: 190,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/5.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.916361,
-      longitude: 6.944974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.7,
-  },
-  {
-    id: 'a6948e68-1886-4b2f-a18b-f1d7ef890051',
-    title: 'Tile House',
-    type: 'apartment',
-    price: 500,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.949361,
-      longitude: 6.976974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.6,
-  },
-  {
-    id: 'fed59ab7-d341-45ca-984a-6fa9f053740e',
-    title: 'Perfectly located Castro',
-    type: 'apartment',
-    price: 410,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/7.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.913361,
-      longitude: 6.9509739999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.1,
-  },
-  {
-    id: '04ed5088-eb0a-4962-8923-2c9580aee03a',
-    title: 'Wood and stone place',
-    type: 'house',
-    price: 449,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.930361,
-      longitude: 6.937974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2,
-  },
-  {
-    id: '4047360b-00f8-4714-850c-7d6115048d25',
-    title: 'Perfectly located Castro',
-    type: 'room',
-    price: 226,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/11.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.960361,
-      longitude: 6.9509739999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.4,
-  },
-  {
-    id: 'db4a0af9-aaf8-47b4-bd04-b82dade39607',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'hotel',
-    price: 192,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.918461,
-      longitude: 6.969974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.6,
-  },
-  {
-    id: '2fe9f28d-fdd6-4795-9150-70dca3ae6ee3',
-    title: 'Wood and stone place',
-    type: 'room',
-    price: 261,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.957361,
-      longitude: 6.9509739999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.1,
-  },
-  {
-    id: '95354afb-7c43-4cee-aeb4-0ebbfd33cd8b',
-    title: 'Waterfront with extraordinary view',
-    type: 'apartment',
-    price: 197,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/6.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.951361,
-      longitude: 6.944974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.2,
-  },
-  {
-    id: '5a1536cb-9b36-419a-ae08-b46d8bcf3e3f',
-    title: 'Tile House',
-    type: 'house',
-    price: 167,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.959361,
-      longitude: 6.978974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.6,
-  },
-  {
-    id: 'b8f330ee-014c-41f9-a7aa-35f786d9f27f',
-    title: 'The Joshua Tree House',
-    type: 'room',
-    price: 150,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.932361,
-      longitude: 6.960974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.4,
-  },
-  {
-    id: '3d0f55af-b1df-4c73-b132-b71de5b88254',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'hotel',
-    price: 327,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.954361,
-      longitude: 6.982974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.9,
-  },
-  {
-    id: 'a176c35b-92b0-4867-a3ff-be99d288955e',
-    title: 'The house among olive ',
-    type: 'room',
-    price: 142,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.945361,
-      longitude: 6.962974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.9,
-  },
-  {
-    id: '00361489-5f2b-4fd6-93be-64bc3f430744',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'house',
-    price: 882,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/6.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.917361,
-      longitude: 6.977974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.5,
-  },
-  {
-    id: '6e2fa971-1b04-4d47-a86a-2c60746d44d0',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'house',
-    price: 574,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/4.jpg',
-    city: {
-      name: 'Cologne',
-      location: {
-        latitude: 50.938361,
-        longitude: 6.959974,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.945361,
-      longitude: 6.935974,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.2,
-  },
-  {
-    id: '683739b7-2cb9-4373-ae70-45072264f9c4',
-    title: 'The Joshua Tree House',
-    type: 'apartment',
-    price: 227,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.854557,
-      longitude: 4.364697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.8,
-  },
-  {
-    id: '6d937b63-e851-43dc-8304-86be6a944990',
-    title: 'Perfectly located Castro',
-    type: 'house',
-    price: 110,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/6.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.867557,
-      longitude: 4.371696999999999,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.3,
-  },
-  {
-    id: '99abcc53-67e5-4c00-a78f-5a68ebc1029e',
-    title: 'House in countryside',
-    type: 'house',
-    price: 818,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/20.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.827557,
-      longitude: 4.336697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.7,
-  },
-  {
-    id: '60365a55-6146-4c87-93d9-36265d79e46b',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'room',
-    price: 187,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.833557,
-      longitude: 4.374696999999999,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.1,
-  },
-  {
-    id: '85c670a4-330b-4a27-afab-2c5002bc0fb3',
-    title: 'The house among olive ',
-    type: 'apartment',
-    price: 243,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.837557,
-      longitude: 4.339697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.2,
-  },
-  {
-    id: 'b54b1685-2d21-46a5-a241-6ec6a179cc48',
-    title: 'Wood and stone place',
-    type: 'house',
-    price: 446,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/4.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.849557,
-      longitude: 4.374696999999999,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.4,
-  },
-  {
-    id: '72640308-c035-4d62-be3e-5250a8d07391',
-    title: 'House in countryside',
-    type: 'house',
-    price: 775,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.852557,
-      longitude: 4.3376969999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4,
-  },
-  {
-    id: '2c7872f7-5e80-4c6f-aed9-37b8d0e98bd8',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'house',
-    price: 979,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.828556999999996,
-      longitude: 4.362697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1,
-  },
-  {
-    id: '17b924cf-d47e-4fc0-a20d-c5c8019a7a61',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'hotel',
-    price: 336,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.835556999999994,
-      longitude: 4.354697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.5,
-  },
-  {
-    id: '249abb2a-2d54-46fd-8e8c-795c8fe794c7',
-    title: 'Penthouse, 4-5 rooms + 5 balconies',
-    type: 'room',
-    price: 226,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/5.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.867557,
-      longitude: 4.339697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.4,
-  },
-  {
-    id: '9fe235d5-4e85-4798-bc2d-6bafc9799239',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'house',
-    price: 216,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.839557,
-      longitude: 4.346697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4,
-  },
-  {
-    id: '9655e76a-ff8c-4643-abbe-6c446f7a5a5a',
-    title: 'Loft Studio in the Central Area',
-    type: 'room',
-    price: 189,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/5.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.844556999999995,
-      longitude: 4.346697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.7,
-  },
-  {
-    id: '288aece3-f104-49c8-9e94-4bc1b30602b1',
-    title: 'Penthouse, 4-5 rooms + 5 balconies',
-    type: 'room',
-    price: 234,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.865556999999995,
-      longitude: 4.371696999999999,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.5,
-  },
-  {
-    id: '30c396f6-2811-4c95-802b-5d9a83e18a36',
-    title: 'House in countryside',
-    type: 'room',
-    price: 194,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.869557,
-      longitude: 4.332697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3,
-  },
-  {
-    id: 'd4748f8f-8116-4c6f-b515-3894e801b85b',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'room',
-    price: 253,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.860557,
-      longitude: 4.376697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.3,
-  },
-  {
-    id: 'fc1d9b03-2731-41f3-a0c2-4d722b4b66fe',
-    title: 'Loft Studio in the Central Area',
-    type: 'house',
-    price: 252,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/9.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.862556999999995,
-      longitude: 4.375697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.4,
-  },
-  {
-    id: 'bb869441-3efa-4dbd-b955-608c79829343',
-    title: 'Waterfront with extraordinary view',
-    type: 'room',
-    price: 129,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/9.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.842557,
-      longitude: 4.3536969999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.8,
-  },
-  {
-    id: '0359d34d-8860-4882-ad91-66474464e039',
-    title: 'Penthouse, 4-5 rooms + 5 balconies',
-    type: 'apartment',
-    price: 410,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/19.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.822556999999996,
-      longitude: 4.347697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.5,
-  },
-  {
-    id: '1981c22c-70fb-422d-ab56-2120cc733205',
-    title: 'Canal View Prinsengracht',
-    type: 'hotel',
-    price: 160,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.867557,
-      longitude: 4.357697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.2,
-  },
-  {
-    id: '97eea6c4-170c-4d91-871e-0aa25aff5642',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'hotel',
-    price: 142,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.842557,
-      longitude: 4.363696999999999,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.5,
-  },
-  {
-    id: '286c107f-5ed9-4a12-ba9e-e16b5cc4836b',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'apartment',
-    price: 263,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36554,
-      longitude: 4.911976,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: true,
-    rating: 1.7,
-  },
-  {
-    id: '00cd8b57-6c63-4b22-8a81-27bdafe72930',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'house',
-    price: 377,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/4.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.385540000000006,
-      longitude: 4.902976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.9,
-  },
-  {
-    id: 'd74c3878-e7bc-499e-accd-2433b188ab19',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'hotel',
-    price: 297,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/20.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.397540000000006,
-      longitude: 4.9099759999999995,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: false,
-    rating: 1.1,
-  },
-  {
-    id: 'd07d21de-2536-4440-8277-1acd837feaa1',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'hotel',
-    price: 138,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/12.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37454,
-      longitude: 4.881976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.7,
-  },
-  {
-    id: '805a9acc-e68c-46c3-9acf-3557ede4838e',
-    title: 'Penthouse, 4-5 rooms + 5 balconies',
-    type: 'apartment',
-    price: 364,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/11.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.367540000000005,
-      longitude: 4.883976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.2,
-  },
-  {
-    id: '020e5a2d-1276-4d97-aec7-d3ad46df522e',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'house',
-    price: 966,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.361540000000005,
-      longitude: 4.883976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.1,
-  },
-  {
-    id: '7e9555a0-8b3d-4baf-aedc-9098291f69e5',
-    title: 'Wood and stone place',
-    type: 'hotel',
-    price: 434,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.35754,
-      longitude: 4.9179759999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.2,
-  },
-  {
-    id: 'af9c36ec-2563-43d2-9cdd-aad7ec8b225a',
-    title: 'House in countryside',
-    type: 'house',
-    price: 464,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/5.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37854,
-      longitude: 4.894976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.7,
-  },
-  {
-    id: '28b76a44-fe2e-43e0-bd3a-51ab0d44ef7f',
-    title: 'House in countryside',
-    type: 'apartment',
-    price: 245,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/7.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36854,
-      longitude: 4.887976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.7,
-  },
-  {
-    id: '31ca02ac-d48d-472c-8256-432be0ff8c9e',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'apartment',
-    price: 127,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/12.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.370540000000005,
-      longitude: 4.9099759999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.1,
-  },
-  {
-    id: '4318c4cb-d6cd-4593-835a-475141387d39',
-    title: 'Tile House',
-    type: 'room',
-    price: 129,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/9.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.389540000000004,
-      longitude: 4.883976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3,
-  },
-  {
-    id: '4dee7261-5e94-4e4d-88ec-ef50320e39aa',
-    title: 'The house among olive ',
-    type: 'house',
-    price: 281,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.35054,
-      longitude: 4.908976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.5,
-  },
-  {
-    id: '0d4eac3c-cb38-421e-97d9-4c2f6f02c56a',
-    title: 'Canal View Prinsengracht',
-    type: 'room',
-    price: 208,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/16.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37154,
-      longitude: 4.889976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.1,
-  },
-  {
-    id: 'f96b6d5a-669a-4bac-bc21-18a11d929e93',
-    title: 'House in countryside',
-    type: 'apartment',
-    price: 126,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.364540000000005,
-      longitude: 4.9019759999999994,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.5,
-  },
-  {
-    id: '45964b10-c604-4906-91c9-b8a02983381f',
-    title: 'Canal View Prinsengracht',
-    type: 'house',
-    price: 797,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36354,
-      longitude: 4.911976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4,
-  },
-  {
-    id: '36930731-3756-40ea-a778-f6d203910772',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'hotel',
-    price: 229,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36354,
-      longitude: 4.889976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.4,
-  },
-  {
-    id: '8700e017-c546-4d25-bd13-7528bf7ef260',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'hotel',
-    price: 415,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/18.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.388540000000006,
-      longitude: 4.899976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.6,
-  },
-  {
-    id: '90b040a7-aa54-4d67-af1a-d208a5e4767d',
-    title: 'Penthouse, 4-5 rooms + 5 balconies',
-    type: 'hotel',
-    price: 114,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36954,
-      longitude: 4.914976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.7,
-  },
-  {
-    id: '89636b36-1053-4492-a2ac-2cda1ebc634d',
-    title: 'Penthouse, 4-5 rooms + 5 balconies',
-    type: 'hotel',
-    price: 207,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/5.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37554,
-      longitude: 4.9019759999999994,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.2,
-  },
-  {
-    id: '005b42c1-5337-4944-9fbf-b7fd9647cd6d',
-    title: 'House in countryside',
-    type: 'room',
-    price: 115,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.385540000000006,
-      longitude: 4.886976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3,
-  },
-  {
-    id: 'f3fc28b3-78a5-429a-9f77-266c191bf62c',
-    title: 'Waterfront with extraordinary view',
-    type: 'hotel',
-    price: 301,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.528341000000005,
-      longitude: 10.018654000000002,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.1,
-  },
-  {
-    id: 'bc5f2339-5208-49e2-a50b-9d9007774cc2',
-    title: 'House in countryside',
-    type: 'room',
-    price: 148,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/12.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.538341,
-      longitude: 9.976654000000002,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.5,
-  },
-  {
-    id: '1d4390fc-4bce-43e4-8fff-3e6b7c7c90cf',
-    title: 'Tile House',
-    type: 'house',
-    price: 147,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/2.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.563341,
-      longitude: 10.019654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.1,
-  },
-  {
-    id: 'ba6341a9-3683-46ab-a802-54ffefda7694',
-    title: 'House in countryside',
-    type: 'apartment',
-    price: 438,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/19.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.529341,
-      longitude: 9.975654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.3,
-  },
-  {
-    id: 'ee652176-cb07-45e8-b267-97a10333a1c9',
-    title: 'Tile House',
-    type: 'apartment',
-    price: 486,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/4.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.565341,
-      longitude: 9.980654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.7,
-  },
-  {
-    id: 'cd590481-49bd-4136-b591-65436cb12695',
-    title: 'Loft Studio in the Central Area',
-    type: 'hotel',
-    price: 173,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/4.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.546341000000005,
-      longitude: 10.022654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.8,
-  },
-  {
-    id: 'a5ad8c5c-677a-424c-953e-8503daaf7af3',
-    title: 'Wood and stone place',
-    type: 'room',
-    price: 204,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/4.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.565341000000004,
-      longitude: 9.978654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.6,
-  },
-  {
-    id: '14c18654-0ec6-486c-9ff9-6cedb4cf9b13',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'apartment',
-    price: 274,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/19.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.570341000000006,
-      longitude: 9.975654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.4,
-  },
-  {
-    id: '7f99e8b7-5bd8-4e92-b825-2f14bc3decff',
-    title: 'The Joshua Tree House',
-    type: 'hotel',
-    price: 228,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.558341000000006,
-      longitude: 10.001654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.4,
-  },
-  {
-    id: '3ee85d5f-9d33-4cad-aa1f-5376e8ca84da',
-    title: 'Perfectly located Castro',
-    type: 'house',
-    price: 905,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.573341000000006,
-      longitude: 10.009654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.4,
-  },
-  {
-    id: '6ae6021f-d43b-476b-a599-7c7c0b892d04',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'apartment',
-    price: 144,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/18.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.573341000000006,
-      longitude: 9.994654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.3,
-  },
-  {
-    id: '04160b6e-76c3-4eb4-93fa-8a0fa64add0b',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'hotel',
-    price: 367,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/20.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.528341000000005,
-      longitude: 9.982654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.9,
-  },
-  {
-    id: '2bae3411-e4a5-449f-bd54-773b2c91ab39',
-    title: 'Wood and stone place',
-    type: 'apartment',
-    price: 126,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/4.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.540341000000005,
-      longitude: 10.025654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1,
-  },
-  {
-    id: '4349a6a7-5ce1-44c6-b1c1-2fc3b8ee40f8',
-    title: 'Canal View Prinsengracht',
-    type: 'apartment',
-    price: 218,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.555341000000006,
-      longitude: 9.975654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.4,
-  },
-  {
-    id: '39125551-34a2-4528-9419-1a27c668cb1c',
-    title: 'Waterfront with extraordinary view',
-    type: 'hotel',
-    price: 102,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.534341000000005,
-      longitude: 9.998654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4,
-  },
-  {
-    id: 'b7da674d-98ae-472c-903f-ddc952ad7ae6',
-    title: 'Waterfront with extraordinary view',
-    type: 'house',
-    price: 207,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.563341,
-      longitude: 9.975654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.6,
-  },
-  {
-    id: 'a1f12bd2-5bbf-413e-8302-b8e2a4d2d466',
-    title: 'Loft Studio in the Central Area',
-    type: 'apartment',
-    price: 262,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/7.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.574341000000004,
-      longitude: 10.022654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.5,
-  },
-  {
-    id: '23fa6167-cf0f-443a-8511-0f639ca138ee',
-    title: 'The Joshua Tree House',
-    type: 'apartment',
-    price: 156,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.573341000000006,
-      longitude: 10.025654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.4,
-  },
-  {
-    id: '54a09b9a-637b-4831-8eae-99b4296d059b',
-    title: 'Waterfront with extraordinary view',
-    type: 'apartment',
-    price: 425,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/5.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.550341,
-      longitude: 9.980654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.5,
-  },
-  {
-    id: '9fcdcc2c-3ca4-4aff-a91e-2a9394e6667e',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'hotel',
-    price: 147,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.558341000000006,
-      longitude: 9.999654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.6,
-  },
-  {
-    id: '09f78347-712c-4a4f-9793-b779ff718d16',
-    title: 'Tile House',
-    type: 'hotel',
-    price: 237,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.236402000000005,
-      longitude: 6.784314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.8,
-  },
-  {
-    id: '0d96167c-8b7a-4fb8-8398-7d0d3e992c8d',
-    title: 'House in countryside',
-    type: 'hotel',
-    price: 169,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/17.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.210402,
-      longitude: 6.798314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.6,
-  },
-  {
-    id: 'cdea1007-c3cb-40cc-b5c9-0af5c0aa14ec',
-    title: 'Wood and stone place',
-    type: 'house',
-    price: 385,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.211402,
-      longitude: 6.756314000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.9,
-  },
-  {
-    id: '59e8488c-0aa1-447a-b545-40267863fe3d',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'house',
-    price: 494,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.228402,
-      longitude: 6.784314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.9,
-  },
-  {
-    id: '5adc2eb1-8e1a-4a34-b698-60573d829b43',
-    title: 'Loft Studio in the Central Area',
-    type: 'hotel',
-    price: 255,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/7.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.204402,
-      longitude: 6.7773140000000005,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.4,
-  },
-  {
-    id: 'd58e7129-fab6-4185-acc2-2288d77f551b',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'room',
-    price: 112,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/20.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.217402,
-      longitude: 6.7693140000000005,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.7,
-  },
-  {
-    id: '52817cfd-1c63-457b-a5bc-b8b2599186e6',
-    title: 'Tile House',
-    type: 'room',
-    price: 130,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.241402,
-      longitude: 6.782314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.5,
-  },
-  {
-    id: '12f42590-af7e-4aa7-89b0-3c6f1124dac3',
-    title: 'Waterfront with extraordinary view',
-    type: 'room',
-    price: 209,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/12.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.235402,
-      longitude: 6.800314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.7,
-  },
-  {
-    id: '0d6ded01-b91b-492a-9882-ec159a048a1f',
-    title: 'Canal View Prinsengracht',
-    type: 'apartment',
-    price: 127,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/15.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.237402,
-      longitude: 6.779314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.8,
-  },
-  {
-    id: '0e46e7d2-3bd8-4f94-bde0-3db4289e760d',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'house',
-    price: 305,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/14.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.237402,
-      longitude: 6.797314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.4,
-  },
-  {
-    id: 'f42bdc80-9c65-4e59-9a81-f4383b4a041f',
-    title: 'House in countryside',
-    type: 'house',
-    price: 293,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.222402,
-      longitude: 6.786314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.3,
-  },
-  {
-    id: 'b908327f-4b7d-415b-ad78-1773755e0605',
-    title: 'Waterfront with extraordinary view',
-    type: 'room',
-    price: 226,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/1.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.232402,
-      longitude: 6.800314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.9,
-  },
-  {
-    id: '961739ab-6a67-4811-a00e-1036097ecaf6',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'apartment',
-    price: 316,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/13.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.216402,
-      longitude: 6.758314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.9,
-  },
-  {
-    id: 'ef7ba464-fc11-4b78-af19-b111e5488a38',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'apartment',
-    price: 374,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.205402,
-      longitude: 6.7613140000000005,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.6,
-  },
-  {
-    id: '006bb444-fa5c-4cf2-bbaf-cd58cb9c8c92',
-    title: 'Waterfront with extraordinary view',
-    type: 'apartment',
-    price: 188,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.248402000000006,
-      longitude: 6.763314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.7,
-  },
-  {
-    id: 'e295da9c-99d0-4418-a4ae-a4db04d273fe',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'apartment',
-    price: 209,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/11.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.250402,
-      longitude: 6.7853140000000005,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.6,
-  },
-  {
-    id: 'a2d09aee-bfe3-4335-9f4d-1519bed3e4cf',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'apartment',
-    price: 389,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.239402000000005,
-      longitude: 6.756314000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.3,
-  },
-  {
-    id: '2f4ed169-0fb7-4770-93c8-464ec606d37c',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'hotel',
-    price: 355,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/3.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.243402,
-      longitude: 6.791314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.5,
-  },
-  {
-    id: '36017de2-a30d-424e-bf26-30ff41d58765',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'house',
-    price: 425,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.214402,
-      longitude: 6.764314000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.6,
-  },
-  {
-    id: '9083425e-e7f5-41f8-9fa4-4ba96f85925d',
-    title: 'Loft Studio in the Central Area',
-    type: 'hotel',
-    price: 355,
-    previewImage: 'https://16.design.htmlacademy.pro/static/hotel/11.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.225402,
-      longitude: 6.784314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.4,
-  },
-];
-
-export const CURRENT_OFFER = {
-  id: '6f6becc6-b05b-4954-a165-6383747770b7',
-  title: 'Nice, cozy, warm big bed apartment',
-  description:
-    'Peaceful studio in the most wanted area in town. Quiet house Near of everything. Completely renovated. Lovely neighbourhood, lot of trendy shops, restaurants and bars in a walking distance.',
-  type: 'house',
-  price: 791,
-  images: [
-    'https://16.design.htmlacademy.pro/static/hotel/8.jpg',
-    'https://16.design.htmlacademy.pro/static/hotel/10.jpg',
-    'https://16.design.htmlacademy.pro/static/hotel/7.jpg',
-    'https://16.design.htmlacademy.pro/static/hotel/18.jpg',
-    'https://16.design.htmlacademy.pro/static/hotel/9.jpg',
-    'https://16.design.htmlacademy.pro/static/hotel/5.jpg',
-  ],
-  city: {
-    name: 'Paris',
-    location: {
-      latitude: 48.85661,
-      longitude: 2.351499,
-      zoom: 13,
-    },
-  },
-  location: {
-    latitude: 48.868610000000004,
-    longitude: 2.342499,
-    zoom: 16,
-  },
-  goods: [
-    'Washer',
-    'Cable TV',
-    'Air conditioning',
-    'Wi-Fi',
-    'Coffee machine',
-    'Laptop friendly',
-    'Towels',
-    'Washing machine',
-    'Heating',
-  ],
-  host: {
-    isPro: true,
-    name: 'Angelina',
-    avatarUrl:
-      'https://16.design.htmlacademy.pro/static/host/avatar-angelina.jpg',
-  },
-  isPremium: true,
-  isFavorite: false,
-  rating: 2.3,
-  bedrooms: 5,
-  maxAdults: 3,
-};
-
-export const getMockNearOfferCardsById = (offerId: string | undefined) =>
-  OFFERS.filter((offer) => offer.id !== offerId).slice(0, 3);
-export const getMockOffer = () => CURRENT_OFFER;
