+++ src/pages/favorites/favorites-page.css
@@ -1,0 +1,3 @@
+.page--favorites {
+  height: 100vh;
+}
